# Backend/gen/api/openapi.yaml
openapi: 3.0.3
info:
  description: Backend API
  title: Backend API
  version: 1.0.0
servers:
- url: https://Backend
paths:
...
  /reviews/{department}/{number}:
    get:
      operationId: getReviews
      parameters:
      - explode: false
        in: path
        name: department
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: number
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: OK
      summary: GET reviews/{department}/{number}
      x-accepts: application/json
...